+++ StreamAbstractionAAMP.h
@@ -1134,7 +1134,7 @@ public:
        * @retval profile index of the current bandwidth
        */
        virtual int GetProfileIndexForBandwidth(long mTsbBandwidth) {
-	       return aamp->mhAbrManager.getBestMatchedProfileIndexByBandWidth(mTsbBandwidth);
+	       return aamp->mhAbrManager.getBestMatchedProfileIndexByBandWidth((int)mTsbBandwidth);
        }
 
 	long GetCurProfIdxBW(){
